timstamp modified
ros::Time( double(pFrameData->timeStamp) * 1.e-4); was giving the time since the duo_node started, but the ros time is different. I was not able to get the pointcloud as the published pointcloud is very old. The above change worked. I was getting pointcloud. I am doubtful about the type of **leftImage.header.stamp** .